requirements.txt
asgiref==3.6.0
Django==4.1.5
django-bootstrap5==22.2
django-cors-headers==3.13.0
django-flatpickr==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
Pillow==9.4.0
pydantic==1.10.4
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.4.0
